Create setting to toggle between windowed and fullscreen.
This required creating an arcade.Window subclass, `GameWindow`, to implement a listener for the `is_fullscreen` config.

Because the screen dimensions change when toggling between fullscreen and windowed, code that referenced the constants SCREEN_WIDTH and SCREEN_HEIGHT was updated to grab the current dimensions.

Name of constants SCREEN_WIDTH and SCREEN_HEIGHT was changed to INITIAL_SCREEN_WIDTH and INITIAL_SCREEN_HEIGHT to reflect the new intended use.